requirements.txt
annotated-types==0.6.0
#asgiref==3.8.1
#cachetools==5.3.3
#certifi==2024.2.2
#charset-normalizer==3.3.2
#colorama==0.4.6
contourpy==1.2.1
#cycler==0.12.1
#dill==0.3.8
dj-rest-auth==6.0.0
# Django==5.0.4
django-cors-headers==4.3.1
django-countries==7.6.1
# django-heroku==0.3.1
#django-js-asset==2.2.0
#django-mptt==0.16.0
djangorestframework==3.15.1
drf-extra-fields==3.7.0
#filelock==3.14.0
filetype==1.2.0
#fonttools==4.51.0
fsspec==2024.5.0
google-ai-generativelanguage==0.6.4
#google-api-core==2.19.0
#google-api-python-client==2.129.0
#google-auth==2.29.0
#google-auth-httplib2==0.2.0
google-generativeai==0.5.4
#googleapis-common-protos==1.63.0
#grpcio==1.63.0
grpcio-status==1.62.2
#httplib2==0.22.0
#idna==3.7
intel-openmp==2021.4.0
#Jinja2==3.1.4
#kiwisolver==1.4.5
#MarkupSafe==2.1.5
# matplotlib==3.9.0
# mkl==2021.4.0
# mpmath==1.3.0
networkx==3.3
# numpy==1.26.4
# opencv-python==4.9.0.80
#packaging==24.0
# pandas==2.2.2
# pillow==10.3.0
proto-plus==1.23.0
#protobuf==4.25.3
#psutil==5.9.8
py-cpuinfo==9.0.0
#pyasn1==0.6.0
#pyasn1_modules==0.4.0
#pydantic==2.7.1
pydantic_core==2.18.2
#pyparsing==3.1.2
#python-dateutil==2.9.0.post0
#pytz==2024.1
#PyYAML==6.0.1
#psycopg2-binary==2.9.9
#requests==2.31.0
#rsa==4.9
# scipy==1.13.0
# seaborn==0.13.2
#six==1.16.0
#sqlparse==0.5.0
# sympy==1.12
tbb==2021.12.0
#tqdm==4.66.4
# torch==1.11.0+cpu
# torchvision==0.12.0+cpu
#typing_extensions==4.11.0
#tzdata==2024.1
ultralytics==8.2.18
#uritemplate==4.1.1
#urllib3==2.2.1
